Reorganized methods into groups.
